Add dataclass for cameras in event processor.
The event processor needs some information about the available cameras.
It needs to know whether or not to ignore events from a given camera.
It is also helpful (but not necessary) for it to have the human-readable
names for the cameras so that logs and filenames are easier to understand.

I could pass in a reference to the ProtectAPIClient, or the cameras object
that it supplies to do this, but I'd prefer to keep the classes more loosely
coupled than that, and even if I did that, I'd have to append the ignore information.

A small dataclass allows me to provide just the camera information that the
event processor needs, and is more readable than passing that data around in dicts.