Convert event processor to module.
It turned out that the event processor didn't need to be a class.
I really only needed the event queue and the list of cameras to persist,
and I could do that with a single closure instead of a full class.

I also got rid of a couple of unneeded private functions while I was at it.