Refactor client manager back to a function.
It also turned out that the client manager didn't need to be a class.
It only provided a single necessary function: the websocker monitor task.
Everything else it did was just holding on to variables or wrapping functions that
were already provided by the ProtectApiClient.

This commit gets rid of all the unnecessary class overhead and moves the monitor
function into the event processor module.