Ensure session is authenticated before attempting to capture video.
The authentication token expires after 59 minutes without activity.
This commit adds a check (actually I had already added it and forgot,
but I moved it into its own function now) just before downloading the
event video to make sure that the session is still authenticated,
and to reauthenticate if it's not.

If the reauthentication fails for some reason, then we log an error
and skip the save attempt, because it's going to fail anyway.